`DataStore` changelog 5: event-driven time histograms (#4208)
This is mostly preliminary work for #4209, which makes this PR a bit
weird. Basically just trying to offload complexity from #4209.

`TimesPerTimeline` as well as `TimeHistogramPerTimeline` are now living
on their own and are maintained as `StoreView`s, i.e. they react to
changes to the `DataStore` rather than constructing alternate truths.

This is the first step towards turning the `EntityTree` giga-structure
into an event-driven view in the next PR.

---

`DataStore` changelog PR series:
- #4202
- #4203
- #4205
- #4206
- #4208
- #4209