Improve input latency by switching to double-buffered rendering
This makes a huge difference in perceived latency on my 60 Hz monitor
on macOS on a native build, e.g. when hovering buttons,
or just the distance between the cursor and the hover-window
when hovering an image.

It does produce some tearing, noticable e.g. when quickly panning
a 2D image. To me the trade-off is absolutely worth it:
tearing is a mild visual glitch that shows up occationally,
input latency is a constant irritation, making an application feel
sluggish.

It makes no difference on web though, so this does NOT
solve https://github.com/rerun-io/rerun/issues/2672