Fix `entity_path_vec!` and `entity_path!` depending on `ToString` being in scope.
The macro used to work in `no_std`+`alloc` conditions, but that was
broken in 0.12 by the addition of the `.to_string()` call. This change
fixes it by making all items used by the macro use fully qualified,
reexported paths; this way, the macro does not depend on the calling
module and crate's imports at all.