Texture support for raw `Mesh3D` logging (#4894)
### What

Meshes can now be logged with optional texture coordinates and an
optional albedo texture!

Vertex texture coordinates are straight forward, implemented as a new
component, mostly to avoid semantic confusion (position2d on mesh?) and
being able to call the coordinates `uv`.

~The (albedo) texture itself got a bit messy: The original plan was to
refer to a separately logged image and I believe that's still where we
want to go. But this messes with some parts of the viewer so I made the
decision to put a `TensorData` based field on the material. The main
drawback of that is that now you can't inspect the texture separately
and the format restrictions are quite tight (this is in part also due to
the mesh renderer itself though, so that's deeper than just not using
the Image archetype). I believe in the future the material should
instead have an entity reference to specify where the texture comes
from.
Another pretty bad sideffect from this is that if you have model with a
bunch of meshes using the same texture several times, it will send this
texture each time to the viewer since there's no way to "instance"
this.~

~... but let's perfect not be the enemy of
well-it-works-for-the-most-important-usecases-so-let's-ship-this!~

Change: Texture is now a component, making a lot of things nicer

---------

Co-authored-by: Clement Rey <cr.rey.clement@gmail.com>