Always show welcome screen, but fade it in some situations
It used to be that `rerun --web-viewer` didn't show the welcome screen,
which is super annoying when testing.
It also feels very brittle to sometimes show it, and sometimes not,
and to keep track if we've ever seen a recording, etc.

SO: the new design is much simpler. We always show the welcome screen,
but if we expect data to come streaming in at any second we delay
it for a a fraction of a second and then fade it in.
This prevents the problem of a flashing welcome screen before showing
actual user data.
This fade-in was proposed during the initial design of the welcome
screen, but we never got to it.

This replaces `--skip-welcome-screen` with `--fade-in-welcome-screen`.

You can test it with `pixi run rerun ----fade-in-welcome-screen`