Always show welcome screen, but fade it in some situations (#5787)
### What
* Closes https://github.com/rerun-io/rerun/issues/4317

It used to be that `rerun --web-viewer` didn't show the welcome screen,
which is super annoying when testing.
It also feels very brittle to sometimes show it, and sometimes not, and
to keep track if we've ever seen a recording, etc.
It also feel inconsistent with how the welcome screen is now just
another app in the recordings panel.

So the new design is much simpler. We always show the welcome screen,
but if we expect data to come streaming in at any moment, we delay the
welcome screen for a a fraction of a second and then fade it in. This
prevents the problem of a flashing welcome screen before showing actual
user data.
This fade-in was proposed during the initial design of the welcome
screen, but we never got to it.

This replaces `--skip-welcome-screen` with `--fade-in-welcome-screen`.

You can test it with `pixi run rerun ----fade-in-welcome-screen`

### Checklist
* [x] I have read and agree to [Contributor
Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and
the [Code of
Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)
* [x] I've included a screenshot or gif (if applicable)
* [x] I have tested the web demo (if applicable):
* Using newly built examples:
[rerun.io/viewer](https://rerun.io/viewer/pr/5787)
* Using examples from latest `main` build:
[rerun.io/viewer](https://rerun.io/viewer/pr/5787?manifest_url=https://app.rerun.io/version/main/examples_manifest.json)
* Using full set of examples from `nightly` build:
[rerun.io/viewer](https://rerun.io/viewer/pr/5787?manifest_url=https://app.rerun.io/version/nightly/examples_manifest.json)
* [x] The PR title and labels are set such as to maximize their
usefulness for the next release's CHANGELOG
* [x] If applicable, add a new check to the [release
checklist](https://github.com/rerun-io/rerun/blob/main/tests/python/release_checklist)!

- [PR Build Summary](https://build.rerun.io/pr/5787)
- [Docs
preview](https://rerun.io/preview/e24d4fd0a23d862697fb1dce865adbd6d1c872b5/docs)
<!--DOCS-PREVIEW-->
- [Examples
preview](https://rerun.io/preview/e24d4fd0a23d862697fb1dce865adbd6d1c872b5/examples)
<!--EXAMPLES-PREVIEW-->
- [Recent benchmark results](https://build.rerun.io/graphs/crates.html)
- [Wasm size tracking](https://build.rerun.io/graphs/sizes.html)

---------

Co-authored-by: Jeremy Leibs <jeremy@rerun.io>