Rewrite `re_ws_comms` to work without `async` & `tokio` runtime (#6005)
### What

* Big chunk of #5907

Removes tokio & async code from re_ws_comms. Achieves this by spawning a
thread for listening and another one for relaying messages (the later
was already the case but via tokio).
Main disadvantage of this change here is that all clients now share the
same thread for sending whereas before this was handled by the tokio
pool. In theory we have less latency though since before we had to go
from the broadcast thread to the client worker threads in a separate
step.

Ended up depending on the
[`polling`](https://docs.rs/polling/latest/polling/index.html) crate
because otherwise shutting down a tcplistener with vanilla std rust is
close to impossible. The version I picked is such that Cargo.lock
doesn't have (yet) another instance of this crate.

This change fixes a bug in the process where we'd not send all so far
received messages to new clients.

⚠️ Rerun serve still needs a tokio runtime with this change ⚠️: To fix
this we need to remove it also for the web server which should be quite
a bit easier 🤞

Testing done:
* [x] `pixi run rerun-web <some rrd>` to serve several local browsers
* [x] `pixi run python .\examples\python\face_tracking\main.py --serve`
to serve several local browsers and checking that they catch up on old
data
* [x] keeping an eye on debug output ensuring that shutdown is graceful
* [x] Test the same on Mac (did only windows so far)

### Checklist
* [x] I have read and agree to [Contributor
Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and
the [Code of
Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)
* [x] I've included a screenshot or gif (if applicable)
* [x] I have tested the web demo (if applicable):
* Using examples from latest `main` build:
[rerun.io/viewer](https://rerun.io/viewer/pr/6005?manifest_url=https://app.rerun.io/version/main/examples_manifest.json)
* Using full set of examples from `nightly` build:
[rerun.io/viewer](https://rerun.io/viewer/pr/6005?manifest_url=https://app.rerun.io/version/nightly/examples_manifest.json)
* [x] The PR title and labels are set such as to maximize their
usefulness for the next release's CHANGELOG
* [x] If applicable, add a new check to the [release
checklist](https://github.com/rerun-io/rerun/blob/main/tests/python/release_checklist)!

- [PR Build Summary](https://build.rerun.io/pr/6005
- [Recent benchmark results](https://build.rerun.io/graphs/crates.html)
- [Wasm size tracking](https://build.rerun.io/graphs/sizes.html)

To run all checks from `main`, comment on the PR with `@rerun-bot
full-check`.