Support shared library building on Windows (#6053)
Developers tend to have strong opinions on shared libraries, that become
even stronger when they see the effort required to get shared libraries
to work on Windows. However, C++ binary package managers (such as
conda/pixi/conda-forge) tend to use shared libraries also on Windows, so
it would be convenient to be able to compile C++ bindings of rerun as
shared libraries on Windows.

This PR does add support for compiling the rerun-cpp library as a shared
library on Windows. The PR is a bit invasive, so feel free to reject it
if you think the maintenance effort is not worth the benefit.

Why all the complexity? In a nutshell, Linux and macOS by default export
all the symbols contained in a shared library. Instead, Windows does not
support that, and require the developers that want to export the symbols
to either annotate every class/attribute/method that needs to be
visible, or pass the attributes with a `.def` file. As this tend to be
cumbersome, CMake has implemented the `WINDOWS_EXPORT_ALL_SYMBOLS`
property, that emulates the "export all symbols" behavior of Linux/macOS
also on Windows. However, it does not support static variables. So for
libraries that do use static variables, can just use
`WINDOWS_EXPORT_ALL_SYMBOLS` and they are good to go. However, if static
variables are used (as in the case of rerun) it is necessary to annotate
every static variable with the appropriate storage-class attribute. The
easiest way to do so is use GenerateExportHeader module to generate a
rerun_sdk_export.hpp header file that define the RERUN_SDK_EXPORT macro,
and add that macro to all static variables. As RERUN_SDK_EXPORT is added
in the C++ source code, we need to make sure that this macro is also
defined in Linux/macOS builds, so the use of generate_export_header is
present even if MSVC AND BUILD_SHARED_LIBS is OFF. This annotation
process is a bit cumbersome in rerun case, as it needs to be done in all
codegen logic that generate `static const` attributes.

### What

### Checklist
* [x] I have read and agree to [Contributor
Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and
the [Code of
Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)
* [x] I've included a screenshot or gif (if applicable)
* [x] I have tested the web demo (if applicable):
* Using examples from latest `main` build:
[rerun.io/viewer](https://rerun.io/viewer/pr/6053?manifest_url=https://app.rerun.io/version/main/examples_manifest.json)
* Using full set of examples from `nightly` build:
[rerun.io/viewer](https://rerun.io/viewer/pr/6053?manifest_url=https://app.rerun.io/version/nightly/examples_manifest.json)
* [x] The PR title and labels are set such as to maximize their
usefulness for the next release's CHANGELOG
* [x] If applicable, add a new check to the [release
checklist](https://github.com/rerun-io/rerun/blob/main/tests/python/release_checklist)!

- [PR Build Summary](https://build.rerun.io/pr/6053)
- [Recent benchmark results](https://build.rerun.io/graphs/crates.html)
- [Wasm size tracking](https://build.rerun.io/graphs/sizes.html)

To run all checks from `main`, comment on the PR with `@rerun-bot
full-check`.

---------

Co-authored-by: Andreas Reich <r_andreas2@web.de>