Reduce number of unwrap calls and make clippy warning for it opt-out per crate (#6311)
<!--
Open the PR up as a draft until you feel it is ready for a proper
review.

Do not make PR:s from your own `main` branch, as that makes it difficult
for reviewers to add their own fixes.

Add any improvements to the branch as new commits to make it easier for
reviewers to follow the progress. All commits will be squashed to a
single commit once the PR is merged into `main`.

Make sure you mention any issues that this PR closes in the description,
as well as any other related issues.

To get an auto-generated PR description you can put "copilot:summary" or
"copilot:walkthrough" anywhere.
-->

### What
Resolves #3408: Remove all usages of unwrap() and forbid it's use (if
applicable).

*If applicable* because even after the first batch of changes there are
over 800(!) usages of unwrap() in the code - for various reasons, and
often explicitly allowed. So I think it's probably unrealistic to get
rid of all of them, and probably also not necessary.

The strategy:
1. Warn on all usages of `unwrap()`.
2. Remove `unwrap()` module by module.
3. Where `unwrap()` would be theoretically okay use `expect("error
msg")`.
4. Leave build tools as they are for now - panics there are annoying but
not critical (can happen when e.g. some command line tool is missing).
5. For tests and benchmarks unwrap is totally okay imho
(`allow-unwrap-in-tests` was anyway already set in `clippy.toml`)
6. Where errors are returned combine error types and "?-ify" the code,
or use some other idiomatic Rust expression.
7. Goal: eventually deny `clippy::unwrap_used` everywhere, unless
explicitly allowed.

*Note: as this is my first PR, early feedback is greatly appreciated*🙏
Also - rookie question - what (or where) are the guidelines for the last
three items in the checklist? Are they needed for this PR, as this is
just a cleanup?

### Checklist
* [x] I have read and agree to [Contributor
Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and
the [Code of
Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)
* [x] I've included a screenshot or gif (if applicable)
* [x] The PR title and labels are set such as to maximize their
usefulness for the next release's CHANGELOG

- [PR Build Summary](https://build.rerun.io/pr/6311)
- [Recent benchmark results](https://build.rerun.io/graphs/crates.html)
- [Wasm size tracking](https://build.rerun.io/graphs/sizes.html)

To run all checks from `main`, comment on the PR with `@rerun-bot
full-check`.