Update pixi.toml to use `rerun_py` package directly (#6466)
### What
Our shim package broke with the update to pixi 0.22.

However, pixi now mostly does the correct thing with a source install of
the `rerun_py` package, so we now use that instead of the shim package.

I introduced a new `python-dev` feature which is used for pulling in the
dev version of rerun. The upside of this is it opens up another feature
variant `examples-pypi` which uses the published version of `rerun`.

We should be able to do something similar with the wheel package as
well.

Lastly, I had to re-generate our lockfile to get around some weird
issues with tzdata. This bumped the lockfile version, which means
dragging everyone forward to pixi 0.23


---------

Co-authored-by: Clement Rey <cr.rey.clement@gmail.com>